+++ trace/src/render/render_graph/LightingPass.cpp
@@ -165,8 +165,8 @@ namespace trace {
 
 				Viewport view_port = renderer->_viewPort;
 				Rect2D rect = renderer->_rect;
-				view_port.width = width;
-				view_port.height = height;
+				view_port.width = static_cast<float>(width);
+				view_port.height = static_cast<float>(height);
 
 				rect.right = width;
 				rect.bottom = height;
